snap: Query VUID for VFs individually
Currently, VUID query for VFs is done in a batch manner for
efficiency, however, if VF number is > 511, the VUIDs returned
exceeds the infiniband DevX commmand limitation, thus will fail
the VF creation.

To solve this problem, based on discussion, this patch changes
the batch query into individual query.

Issue: 3365502
Signed-off-by: Wei Zhang <weizhang@nvidia.com>