virtio/aq: Refactor admin virtqueue processing for different versions
Refactor the version control code in the admin virtqueue to improve
maintainability and reliability. Previously, the code used if statements
to handle different virtio spec versions. With this refactor, the code
adopts the concept of interface-based object orientation and utilizes
function pointers as the operation interface. As a result, the version
number judgments are eliminated, allowing the admin virtqueue processing
function to adapt seamlessly to different virtio spec versions.

Issue: 3404271
Signed-off-by: Feng Liu <feliu@nvidia.com>