virtio_blk: create admin q on core #0
Related to #3585379.

Based on a discussion with Alex:
the response to quiesce (snap_virtio_ctrl_quiesce_adm_done)
can be done in the context of the 'admin' poller.
Note: the vblk admin queue may be scheduled on the core
that differs from the core of the admin poller. As a result,
it might cause multithreaded access to the dma queue.

To prevent the access to admin queue from different
threads - create an admin queue on core #0.

Signed-off-by: Rostislav Raikhman <rraikhman@nvidia.com>