virtio-blk: fix #3585379
The issue is the following:

Snap receives the admin command, processes it,
and sends the completion to the host. Because of the bug
the ring of the doorbell (the RING_BATCH is the default),
which should be applied in the next iteration of
The progress is not applied.

As a result, the completion is not handled by the host.
If at this point, the dbg_stat RPC command is sent
to snap, the processing is resumed!

The issue is resolved by db_stats RPC commands because the
processing of the RPC command initiates the 'flush' on the
admin queue. The flush does the progress of DMA tx
and rings the doorbell.

The issue was raised because after restarting the snap
the admin queue was not bound to the same thread as
the 'admin poller thread'.

The change 19d317b5ed83564dfa151759c1438c783be30dd8
fixed the issue only upon startup of snap when the
driver not loaded.

Fix:
Bound admin queue into 'admin poller' core
('polling group 0') when the controller is under recovery
(resumed).

Signed-off-by: Rostislav Raikhman <rraikhman@nvidia.com>