snap_dma: Fix max WR count check at snap_dma_build_sgl
When the number of WR equals to the maximum value, the function fails.
The failure should happen only when the number of WR is higher than
the maximum value. The commit fixes it by moving the check to the
start of the loop.
Also, reduce the debug level of prints when QP is out of tx_available
resource, since the requests go to the pending list.

Signed-off-by: Israel Rukshin <israelr@nvidia.com>