snap_dma: error recovery API
The commit introduces API that can be used to add
custom error handling, migrate uncompleted operations to a new
QP and to reconnect failed endpoints.

Main purpose of the API is to allow recovery after remote access
error that was triggered by the FLR.

This is a necessary step for 'single QP per core' designs.

Signed-off-by: Alex Mikheev <amikheev@nvidia.com>